Make ModusDataPacket an inner class of CaptchaDeckPacket
