+++ src/main/java/com/mraof/minestuck/inventory/captchalogue/CaptchaDeckHandler.java
@@ -7,7 +7,7 @@ import com.mraof.minestuck.alchemy.AlchemyHelper;
 import com.mraof.minestuck.computer.editmode.ServerEditHandler;
 import com.mraof.minestuck.item.BoondollarsItem;
 import com.mraof.minestuck.item.MSItems;
-import com.mraof.minestuck.network.data.ModusDataPacket;
+import com.mraof.minestuck.network.CaptchaDeckPacket;
 import com.mraof.minestuck.player.ClientPlayerData;
 import com.mraof.minestuck.player.PlayerBoondollars;
 import com.mraof.minestuck.player.PlayerData;
@@ -62,7 +62,7 @@ public final class CaptchaDeckHandler
 		
 		ModusHolder modusHolder = getHolder(player);
 		if(modusHolder.modus != null)
-			player.connection.send(ModusDataPacket.create(modusHolder.modus));
+			player.connection.send(CaptchaDeckPacket.ModusData.create(modusHolder.modus));
 		
 		if(modusHolder.modus == null && !modusHolder.givenModus)
 			CaptchaDeckHandler.tryGiveStartingModus(modusHolder, player);
@@ -345,7 +345,7 @@ public final class CaptchaDeckHandler
 		} else
 		{
 			modus.initModus(null, player, null, size);
-			player.connection.send(ModusDataPacket.create(modus));
+			player.connection.send(CaptchaDeckPacket.ModusData.create(modus));
 		}
 	}
 	
@@ -419,7 +419,7 @@ public final class CaptchaDeckHandler
 		modusHolder.modus = modus;
 		if(modus != null)
 			modusHolder.givenModus = true;
-		player.connection.send(ModusDataPacket.create(modus));
+		player.connection.send(CaptchaDeckPacket.ModusData.create(modus));
 	}
 	
 	private static void tryGiveStartingModus(ModusHolder modusHolder, ServerPlayer player)
@@ -486,3 +486,4 @@ public final class CaptchaDeckHandler
 		}
 	}
 }
+
