+++ src/main/java/com/mraof/minestuck/inventory/captchalogue/Modus.java
@@ -1,6 +1,6 @@
 package com.mraof.minestuck.inventory.captchalogue;
 
-import com.mraof.minestuck.network.data.ModusDataPacket;
+import com.mraof.minestuck.network.CaptchaDeckPacket;
 import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
@@ -84,7 +84,7 @@ public abstract class Modus
 	{
 		if(needResend)
 		{
-			PacketDistributor.PLAYER.with(player).send(ModusDataPacket.create(this));
+			PacketDistributor.PLAYER.with(player).send(CaptchaDeckPacket.ModusData.create(this));
 			needResend = false;
 		}
 	}
